Regexp matched empty pids, which happened on empty lines
the exception thrown on any such line impedes collection of
"good" javaPids, which makes extractPid fail as a whole
instead of returning a list of pids.

For tomcat, the line
    Process proc3 = ['grep', '.sh'].execute()
will fail always, because tomcat startup spawns java and thus
there is not .sh on the ps list. So collecting java processes
is the only working code path. And this one was failing in some cases.